- comment out and remove some dead code from app.js - create a folder for EditProject and its subcomponents --subject to change-- - attempt to clean up conditionals by defining conditional values at once instead of individually --subject to change-- - clean up prop passing in EditProject to avoid passing down unnecessary props --incomplete, have a good feeling the designer/engineer count may also be relocated-- - begin to work on tool-rendering bug where tools part of a project may be added to the project twice --this is due to projectTools and allTools being different objects in memory, therefore failing .includes method-- - subtle renaming in EditProject/helpers for semanticism
